Wire: allow changing buffer limits
Previously, all TwoWire instances share the same RX and TX buffers,
and buffer limits are hard-coded, making it impossible to interact
with I2C devices that require a larger buffer.

This commit moves RX and TX buffers as member variable of TwoWire
class, and introduces a Wire.changeBufferLimits(rxLimit, txLimit)
function that allows customizing RX and TX buffer limits.

Also, the redundant txBufferIndex variable is removed in favor of
txBufferLength.